Move Permissions test cases
