+++ tests/protocols/gmpv208/entities/permissions/test_clone_permission.py
@@ -16,27 +16,20 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-from gvm.errors import GvmError
+from gvm.errors import RequiredArgument
 
 
-class GmpDeletePermissionTestCase:
-    def test_delete(self):
-        self.gmp.delete_permission('a1')
-
-        self.connection.send.has_been_called_with(
-            '<delete_permission permission_id="a1" ultimate="0"/>'
-        )
-
-    def test_delete_ultimate(self):
-        self.gmp.delete_permission('a1', ultimate=True)
+class GmpClonePermissionTestMixin:
+    def test_clone(self):
+        self.gmp.clone_permission('a1')
 
         self.connection.send.has_been_called_with(
-            '<delete_permission permission_id="a1" ultimate="1"/>'
+            '<create_permission>' '<copy>a1</copy>' '</create_permission>'
         )
 
     def test_missing_id(self):
-        with self.assertRaises(GvmError):
-            self.gmp.delete_permission(None)
+        with self.assertRaises(RequiredArgument):
+            self.gmp.clone_permission('')
 
-        with self.assertRaises(GvmError):
-            self.gmp.delete_permission('')
+        with self.assertRaises(RequiredArgument):
+            self.gmp.clone_permission(None)
