+++ tests/protocols/gmpv208/testcmds/__init__.py
@@ -19,11 +19,9 @@
 # pylint: disable=no-member
 
 from .test_create_filter import GmpCreateFilterTestCase
-from .test_create_permission import GmpCreatePermissionTestCase
 from .test_create_tag import GmpCreateTagTestCase
 from .test_get_aggregates import GmpGetAggregatesTestCase
 from .test_modify_filter import GmpModifyFilterTestCase
-from .test_modify_permission import GmpModifyPermissionTestCase
 from .test_modify_report_format import GmpModifyReportFormatTestCase
 from .test_modify_tag import GmpModifyTagTestCase
 from .test_get_feed import GmpGetFeedTestCase
@@ -42,22 +40,17 @@ from .test_authenticate import GmpAuthenticateTestCase
 from .test_clone_report_format import GmpCloneReportFormatTestCase
 from .test_clone_filter import GmpCloneFilterTestCase
 from .test_clone_group import GmpCloneGroupTestCase
-from .test_clone_permission import GmpClonePermissionTestCase
 from .test_clone_role import GmpCloneRoleTestCase
 from .test_clone_schedule import GmpCloneScheduleTestCase
 from .test_clone_tag import GmpCloneTagTestCase
-from .test_clone_user import GmpCloneUserTestCase
 from .test_create_group import GmpCreateGroupTestCase
 from .test_create_role import GmpCreateRoleTestCase
-from .test_create_user import GmpCreateUserTestCase
 from .test_delete_filter import GmpDeleteFilterTestCase
 from .test_delete_group import GmpDeleteGroupTestCase
-from .test_delete_permission import GmpDeletePermissionTestCase
 from .test_delete_report_format import GmpDeleteReportFormatTestCase
 from .test_delete_role import GmpDeleteRoleTestCase
 from .test_delete_schedule import GmpDeleteScheduleTestCase
 from .test_delete_tag import GmpDeleteTagTestCase
-from .test_delete_user import GmpDeleteUserTestCase
 from .test_describe_auth import GmpDescribeAuthCommandTestCase
 from .test_empty_trashcan import GmpEmptyTrashcanCommandTestCase
 from .test_get_feeds import GmpGetFeedsTestCase
@@ -65,8 +58,6 @@ from .test_get_filter import GmpGetFilterTestCase
 from .test_get_filters import GmpGetFiltersTestCase
 from .test_get_group import GmpGetGroupTestCase
 from .test_get_groups import GmpGetGroupsTestCase
-from .test_get_permission import GmpGetPermissionTestCase
-from .test_get_permissions import GmpGetPermissionsTestCase
 from .test_get_preference import GmpGetPreferenceTestCase
 from .test_get_preferences import GmpGetPreferencesTestCase
 from .test_get_report_format import GmpGetReportFormatTestCase
@@ -80,8 +71,6 @@ from .test_get_settings import GmpGetSettingsTestCase
 from .test_get_system_reports import GmpGetSystemReportsTestCase
 from .test_get_tag import GmpGetTagTestCase
 from .test_get_tags import GmpGetTagsTestCase
-from .test_get_user import GmpGetUserTestCase
-from .test_get_users import GmpGetUsersTestCase
 from .test_get_version import GmpGetVersionCommandTestCase
 from .test_help import GmpHelpTestCase
 from .test_import_report_format import GmpImportReportFormatTestCase
@@ -89,7 +78,6 @@ from .test_modify_auth import GmpModifyAuthTestCase
 from .test_modify_group import GmpModifyGroupTestCase
 from .test_modify_role import GmpModifyRoleTestCase
 from .test_modify_setting import GmpModifySettingTestCase
-from .test_modify_user import GmpModifyUserTestCase
 from .test_restore import GmpRestoreTestCase
 from .test_sync_cert import GmpSyncCertCommandTestCase
 from .test_sync_feed import GmpSyncFeedCommandTestCase
